+++ apps/web/app/page.tsx
@@ -59,7 +59,7 @@ export default function Page(): JSX.Element {
           <div className="flex bg-zinc-900 sticky top-0 z-20 min-w-fit">
             <RunColumnHeaders
               showPrompt={(run: RunCompletion) =>
-                showRunDetails(activeRun ? undefined : run)
+                showRunDetails(activeRun?.id === run.id ? undefined : run)
               }
               headers={tableHeaders}
             />