+++ packages/types/src/index.ts
@@ -45,27 +45,58 @@ interface RunConfigBase {
   type: string;
   name?: string;
   scorers?: Scorer[];
+  parameters?: {
+    [key: string]: any;
+  };
+}
+
+type ResponseFormat = {
+  type: "json_object" | "text";
+};
+
+interface ModelParameters {
+  // From OpenAI config: see OpenAI.ChatCompletionCreateParamsNonStreaming
+  // TODO: This does not support logit_bias, tools, tool_choice, user, stream
+  temperature?: number;
+  max_tokens?: number;
+  top_p?: number;
+  frequency_penalty?: number;
+  logprobs?: boolean;
+  n?: number;
+  presence_penalty?: number;
+  response_format?: ResponseFormat;
+  seed?: number;
+  stop?: string | Array<string>;
+  top_logprobs?: number;
+
+  // For other models, we coerce the above known parameters to appropriate slots
+  // If users require other parameters, we support passthrough for other key names
+  [key: string]: any;
 }
 
 export interface ModelRunConfig extends RunConfigBase {
   type: "model";
-  provider: "openai" | "mistral" | "google" | "anthropic";
+  provider: "openai" | "mistral" | "google" | "anthropic" | "fireworks";
   model: string;
   prompt?: Prompt;
+  parameters?: ModelParameters;
 }
 
 export interface JSScriptRunConfig extends RunConfigBase {
   type: "js-script";
-  value: string;
+  path: string;
 }
 
 export interface PyScriptRunConfig extends RunConfigBase {
   type: "py-script";
-  value: string;
-  pythonPath?: string;
+  path: string;
+  parameters?: {
+    pythonPath?: string;
+    [key: string]: any;
+  };
 }
 
-export type IRunConfig = ModelRunConfig | PyScriptRunConfig | JSScriptRunConfig;
+export type RunConfig = ModelRunConfig | PyScriptRunConfig | JSScriptRunConfig;
 
 export interface ScoreStats {
   name: string;
@@ -84,7 +115,7 @@ export interface RunOutputStats {
 
 export interface RunCompletion {
   id: string;
-  run_config: IRunConfig;
+  run_config: RunConfig;
   dataset_config: { id: string };
   samples: RunOutputSample[];
   stats?: RunOutputStats;